image cached filenames implimented
for speed, every image generated will be stored as part of this storage
each image will need a unique name. A method was implimented that takes
the input parameters from the HTTP GET request and makes a filename
using them. The GET request is key:value in nature, all keys and
parameters are forced to lowercase and alphabetized so that the same
inputs in any order will result in the same filename