Continued work on the interpretation of HTTP GET
The way for dealing with the output will be a seperate method for each
output that will handle the various calculations. The methods that
interpret the GET request have been modified to raise ValueError when
the GET request contains something that isn't handled by the model in
the production server it should display some generic error image.

Things that should be done:
Extensively document the crust pickler
document everything!