Isolate basemap with multitasking
I've done it! It took a bunch of time reading the various documentations
about best practices with how to isolate the memory for each of these
basemap instances (pyplot was the real unsafe thing). I finally found
that multitasking was the way to go as I'm not IO bound (I don't need to
block). This also means I should get a real good speed boost on systems
with more than one processor. I've also adopted the practice of keeping
any and all methods that will not be called by the pyramid routes out of
the views.py file.