changed internal data of crust_model
The data is now one large table that is made on init. There is one
column per derived parameter and each method writes to these unique
columns. Also the griddata data column is selected by the select_output
method now and can use any column for output.

Also added was some documentation. Things should be ready for the
implimentation of some heat flow calcualtions. To be done is what will
sum up the inverse suqare contribution of the geoneutrinos.